+++ CHANGELOG.md
@@ -46,6 +46,9 @@ and this project does not adhere to [Semantic Versioning](http://semver.org/).
   spaces possible, the interfaces of those classes now take `shared_ptr`s
   to tuples of spaces. To create such `shared_ptr`s to tuples, you can use
   the new function `make_space_tuple`.
+* the example programs plot_solution and dune_dpg_error now require
+  dune-subgrid to compile as they use `HangingNodeP2NodalBasis` for the
+  trace variables now.
 * `BoundaryTools::getInflowBoundaryValue` has been renamed to
   `BoundaryTools::getBoundaryValue` as it also works for other boundary
   values. Additionaly, it now takes a single boundary value function